refactor: 불필요한 코드 삭제 및 의존성 수정
