+++ src/test/java/com/idea5/four_cut_photos_map/domain/review/service/GetReviewServiceImplTest.java
@@ -12,7 +12,7 @@ import com.idea5.four_cut_photos_map.domain.review.entity.Review;
 import com.idea5.four_cut_photos_map.domain.review.entity.enums.ItemScore;
 import com.idea5.four_cut_photos_map.domain.review.entity.enums.PurityScore;
 import com.idea5.four_cut_photos_map.domain.review.entity.enums.RetouchScore;
-import com.idea5.four_cut_photos_map.domain.review.mapper.ReviewMapper2;
+import com.idea5.four_cut_photos_map.domain.review.mapper.ReviewMapper;
 import com.idea5.four_cut_photos_map.domain.review.repository.ReviewRepository;
 import com.idea5.four_cut_photos_map.domain.shop.dto.response.ShopResponse;
 import com.idea5.four_cut_photos_map.domain.shop.entity.Shop;
@@ -45,7 +45,7 @@ public class GetReviewServiceImplTest {
     private MemberTitleLogRepository memberTitleLogRepository;
 
     @Mock
-    private ReviewMapper2 reviewMapper;
+    private ReviewMapper reviewMapper;
     @Mock
     private MemberMapper memberMapper;
     @Mock
